Fix for clang

+++ cmd/dbtool.cpp
@@ -878,7 +878,7 @@ int main(int argc, char* argv[]) {
             std::cout << " Provided --commit size is invalid" << std::endl;
             return -1;
         }
-        copy_opts.commitsize = std::max(*tmpsize, 1ull << 20);
+        copy_opts.commitsize = std::max(*tmpsize, static_cast<uint64_t>(1ull << 20));
         tmpsize.reset();
     }
